mais um teste adicionado

+++ sources/ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: gabriel <gabriel@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/02 11:16:02 by gabriel           #+#    #+#             */
-/*   Updated: 2021/02/03 14:02:59 by gabriel          ###   ########.fr       */
+/*   Updated: 2021/02/11 13:32:11 by gabriel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int	contador;
+	size_t	contador;
 
 	contador = 0;
 	while (contador < n)
