+++ sources/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: gabriel <gabriel@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/03 08:48:24 by gabriel           #+#    #+#             */
-/*   Updated: 2021/02/04 11:15:43 by gabriel          ###   ########.fr       */
+/*   Updated: 2021/02/11 13:37:51 by gabriel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int				contador;
+	size_t			contador;
 	unsigned char	left;
 	unsigned char	right;
