+++ sources/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: gabriel <gabriel@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/02 14:17:00 by gabriel           #+#    #+#             */
-/*   Updated: 2021/02/08 12:03:02 by gabriel          ###   ########.fr       */
+/*   Updated: 2021/02/11 13:37:14 by gabriel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int	contador;
+	size_t	contador;
 
 	if (dest == NULL && src == NULL)
 		return (NULL);
