+++ sources/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: gabriel <gabriel@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/02 09:10:12 by gabriel           #+#    #+#             */
-/*   Updated: 2021/02/09 17:24:08 by gabriel          ###   ########.fr       */
+/*   Updated: 2021/02/11 13:31:34 by gabriel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	int	contador;
+	size_t	contador;
 
 	contador = 0;
 	while (contador < n)
